* develop:
  Proper root logging setting for spring-boot 1.3.x
  fix missing license
  Added missing license header
  Fixed typo in field name
  Added indexes for involved search query's fields
  Added missing test files
  Fit iam-robot-testsuite tests and cosmetic fixes
  Fixed missing docker-compose trust service volume
  Fixed Sonar issue
  Removed old Users and Groups pages
  Added Users2 and Groups2 web pages
  Improved User and Group search API

+++ iam-login-service/src/main/java/it/infn/mw/iam/core/group/IamGroupService.java
@@ -24,7 +24,7 @@ import it.infn.mw.iam.persistence.model.IamGroup;
 
 public interface IamGroupService {
   
-  int countAllGroups();
+  long countAllGroups();
   Page<IamGroup> findAll(Pageable page);
   
   Optional<IamGroup> findByNameWithDifferentId(String name, String uuid);
